requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
yfinance==0.2.36
scikit-learn==1.3.2
plotly==5.18.0
ta==0.10.2 